int is faster than int8

+++ enc_functs.py
@@ -59,7 +59,7 @@ def OneSectOctMask2( icPC,  BB007, BB007M, SectSize, StartStopLength, StartStopL
     # % StartStopLengthM=StartStopLengthM(1:inds(2),:);
 
 # %% 0. Mark the TRUE points on BWTrue
-    BWTrue = np.zeros( (SectSize[0], SectSize[1]),'int8')
+    BWTrue = np.zeros( (SectSize[0], SectSize[1]),'int')
     for iBB in range(StartStopLength[icPC,0],StartStopLength[icPC,1]+1):
 
         BWTrue[BB007[iBB,2], BB007[iBB,0]] = 1
@@ -69,7 +69,7 @@ def OneSectOctMask2( icPC,  BB007, BB007M, SectSize, StartStopLength, StartStopL
 
 # %% 0.1 Mark the PREVIOUS SECTION TRUE points on BWTrue
 
-    BWTrue1 = np.zeros( (SectSize[0], SectSize[1] ),'int8')
+    BWTrue1 = np.zeros( (SectSize[0], SectSize[1] ),'int')
     if(icPC > 0):
         if( StartStopLength[icPC-1,1] > 0 ):
             for iBB in range(StartStopLength[icPC-1,0],StartStopLength[icPC-1,1]+1):
@@ -78,7 +78,7 @@ def OneSectOctMask2( icPC,  BB007, BB007M, SectSize, StartStopLength, StartStopL
 
 # %% 0.2 Mark the PREVIOUS_PREVIOUS SECTION TRUE points on BWTrue
 
-    BWTrue2 = np.zeros( (SectSize[0], SectSize[1]),'int8')
+    BWTrue2 = np.zeros( (SectSize[0], SectSize[1]),'int')
     if(icPC > 1):
         if( StartStopLength[icPC-2,0] > 0 ):
             for iBB in range(StartStopLength[icPC-2,0],StartStopLength[icPC-2,1]+1):
@@ -95,15 +95,15 @@ def OneSectOctMask2( icPC,  BB007, BB007M, SectSize, StartStopLength, StartStopL
     # % Go over the possible points
     iTemp = -1
     TempCaus = [Tempaxa[0][0:TCsize],Tempaxa[1][0:TCsize]]
-    disp0 = np.zeros((TCsize,),'int8')
-    disp1 = np.zeros((TCsize,),'int8')
+    disp0 = np.zeros((TCsize,),'int')
+    disp1 = np.zeros((TCsize,),'int')
     for i1 in range( TCsize):
 
         disp0[i1] = TempCaus[0][i1]
         disp1[i1] = TempCaus[1][i1]
         
-    Temp = np.zeros((StartStopLengthM[icPC,2],ctx_type),'int8')
-    Des = np.zeros((StartStopLengthM[icPC,2],),'int8')
+    Temp = np.zeros((StartStopLengthM[icPC,2],ctx_type),'int')
+    Des = np.zeros((StartStopLengthM[icPC,2],),'int')
 
     
     
@@ -115,7 +115,7 @@ def OneSectOctMask2( icPC,  BB007, BB007M, SectSize, StartStopLength, StartStopL
         Temp2 = BWTrue2[(ir-b):(ir+b+1),(ic-b):(ic+b+1)].flatten('F') #(2b+1)**2
         Temp1 = BWTrue1[(ir-b):(ir+b+1),(ic-b):(ic+b+1)].flatten('F')
     
-        TCaus = np.zeros((TCsize,),'int8')
+        TCaus = np.zeros((TCsize,),'int')
         for i1 in range( TCsize):
             # try:
             # TCaus[i1] = BWTrue[ir+TempCaus[0][i1]+pad, ic+TempCaus[1][i1]+pad]
@@ -208,8 +208,8 @@ def get_temps_dests2(BB007,ctx_type):
     # %%
     nM = np.max(BB007M[:,1])
     
-    Temps = np.zeros((BB007M.shape[0],ctx_type),'int8')
-    Dests = np.zeros((BB007M.shape[0]),'int8')
+    Temps = np.zeros((BB007M.shape[0],ctx_type),'int')
+    Dests = np.zeros((BB007M.shape[0]),'int')
